resolved circular deps
